requirements.txt
mcp
google-cloud-storage
python-dotenv
uvicorn
starlette
google-genai
